Adding statistics in APIGeoloc component.
