--- front/src/components/css/GeoStatistics.css
@@ -1,7 +1,0 @@
-.stats {
-  font-size: 2em;
-}
-
-.statistics {
-  text-align: left;
-}
